Ref #7
Add recent plot hist changes.

+++ util/Conventional_Monitor/image_gen/ush/read_scatter.sh
@@ -4,19 +4,13 @@ set -xa
 #
 #  read_scatter.sh
 #
+#    Extract a subset of data from the scater file
+#    and generate an out_${mtype) data file and 
+#    control file for GrADS.
 #-----------------------------------------------------------------------------------
 
 echo "---> read_scatter.sh"
 echo " CONMON_SUFFIX = $CONMON_SUFFIX"
-
-##if  [ $# -ne 1 ] ; then
-## echo "usage: $0 date"
-## exit 8
-##fi
-
-
-## Set mydir.  Remove and make clean mydir.  cd mydir
-##????
 
 exp=$1
 echo "exp set to $exp"
@@ -43,24 +37,12 @@ echo "datadir set to $datadir"
 sorcdir=${12}
 echo "sorcdir set to $sorcdir"
 
-##exp=copr
-## dtype=uv220_00
-## mtype=uv220
-##subtype=00
-## rdate=2007071100
-## fixdir=/nwprod/fix
-## nreal=21
-## exec=read_uv.x
-## type=uv
-## cycle=ges
-## datadir=/u/wx20es/nbns/stats/convweb/copr/horz_hist/ges
-## sorcdir=/u/wx20es/home/convweb/exec
-
-
 ## set up the directory with excutable files
 
 fixfile=global_convinfo.txt 
-cp ${fixdir}/${fixfile} ./convinfo
+if [[ ! -e ./convinfo ]]; then
+   cp ${fixdir}/${fixfile} ./convinfo
+fi
 
 fname=$datadir/${dtype}.scater.${cycle}.${rdate}
 
@@ -73,20 +55,22 @@ fname=$datadir/${dtype}.scater.${cycle}.${rdate}
 rm -f input
 cat << EOF > input
   &input 
-  nreal=${nreal},mtype='${mtype}',fname='${fname}',fileo='out',rlev=0.1,insubtype=${subtype},
+  nreal=${nreal},
+  mtype='${mtype}',
+  fname='${fname}',
+  fileo='out_${dtype}_${cycle}.${rdate}',
+  rlev=0.1,
+  insubtype=${subtype},
 /
 EOF
 
 cp $sorcdir/$exec ./$exec
 
-./$exec <input  > stdout  2>&1
+./$exec <input  > stdout_${dtype}_${cycle}.${rdate}  2>&1
 
 #rm -f $exec
 #rm -f input
-#rm -f fname.out
 
-mv out out_${dtype}_${cycle}.${rdate}
-mv stdout stdout_${dtype}_${cycle}.${rdate}
 
 if [ "${type}" = 'uv' ]; then
    mv out_u out_${dtype}_u_${cycle}.${rdate}
